Add config to store SCM revision in manifest
This uses the buildnumber-maven-plugin, which is quite nice.

To use this capability, it is still necessary to add
buildnumber-maven-plugin to your project POM's <plugins> section, as
well as to define the <scm> section correctly.

Note that while the MANIFEST.MF in the JAR will be correct, the one in
target/classes/META-INF may be partial or missing.

For more information, see:
    http://mojo.codehaus.org/buildnumber-maven-plugin/usage.html