Add a project dedicated to mirror the ImageJ 1.x website
It might come in handy to mirror other static websites, too.

Originally, this was a Beanshell script that was used on fiji.sc (but
hung too often and was replaced by an rsync script anyway, for a while).
It has been converted into a real Java class here and acquired
command-line options, too.

Since imagej.nih.gov locks us out if we request too many things too
quickly, we needed to add code to wait a little after every request.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>