Add the imagej-maven-plugin
For ImageJ 1.x plugins, we want the convenience that the plugin and its
dependencies are copied to ImageJ.app/plugins/ and ImageJ.app/jars/,
respectively.

This plugin's purpose is exactly that.

It provides the 'copy-jars' goal which is bound to the 'install' life
cycle phase by default.

The required parameter 'imagejDirectory' is expected to point to a
valid ImageJ.app/ directory structure.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>