imagej-maven-plugin: do not force-copy the top-level artifact
The idea was that we always want to copy the top-level artifact, but
not necessarily the dependencies.

But since Maven visits all the dependencies in src-plugins/ anyway,
the desired effect cannot be achieved.

In any case, it might make much more sense to copy artifacts only
when the corresponding file in ImageJ.app/ does not exist or is older,
anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>