imagej-maven-plugin: switch from an explicit parameter to an implicit one
Earlier, we required the explicit configuration parameter
'imagejDirectory' for the 'copy-jars' goal. However, this requires
everybody who wants to use this goal to specify the configuration
explicitly at least in the root POM, which in turn requires the whole
<build><plugins><plugin>...</plugin></plugins></build> dance.

Instead, we switched to a parameter called 'imagejDirectoryProperty'
which refers to the name of a property specifying the location of the
ImageJ.app/ directory, and we skip copying anything if that property
is not set or does not point to a valid directory.

That way, we can add this plugin and its two goals to the default
life cycle incurred by using the SciJava POM; developers who want to
install their plugin to their ImageJ.app/ directory can then specify
the location either explicitly in the pom.xml file, or conveniently
via the command-line, depending whether their ImageJ.app/ directory
is expected to be in the project directory structure or outside.

Example: assume you developed your own plugin outside of Fiji or
ImageJ2. Then you would build the plugin in your favorite development
environment, and once you want to upload it to your update site,
you would call something like:

	mvn -Dimagej.app.directory=/home/gene099/Desktop/Fiji.app/

to install the plugin and its dependencies into your Fiji directory.
After that is done, you can start Fiji and upload your plugin; the
update will ask you to upload any used dependencies so that your
users will install and update all the required libraries as well.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>