imagej-maven-plugin: allow the ImageJ.app to be set on the command line
It is really convenient to call

	mvn -Dimagej.app.directory=/home/gene099/Desktop/Fiji.app

to have the artifact and dependencies be copied to the given directory,
without having to modify any configuration or asking for a specific
goal.

If both the pom.xml and the command line specify the property, the latter
takes precedence (to allow convenient overriding via the command line).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>