Use plexus-interpolation in imagej-maven-plugin
It seems really overly complex to add yet another dependency, not even
a core Maven one, to perform a task that should be pretty central to
Maven's execution: interpolation of expressions. But there you go,
it *is* that complex.

This change is necessary to handle both Maven 2 and Maven 3 correctly.
Previously, this plugin only worked with Maven 2 since Maven 3 no longer
offers the MavenProject#getProjectBuilderConfiguartion() method.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>