Add a deploy helper
When running 'mvn deploy', Maven thinks it is okay to re-deploy
non-SNAPSHOT versions.

Even with -SNAPSHOT versions, it is not helpful to deploy unchanged
artifacts, because everybody who builds dependent projects will have
to download 'new' versions unnecessarily.

This helper script scans a Git checkout for artifacts that really
have changed, and deploys only those.

This shell script has been developed in the context of Fiji originally,
but it is too useful to be confined to that project.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>