Do not declare a default type for plugins
This is important, because it is too easy to leave off the type and have
it end up categorized as a generic SciJavaPlugin, when really it should
be something else.

Previously, ImageJ used a default type of Command, but since the Command
package remained in ImageJ, it cannot be the default for all of SciJava.

It is safer to require people to always declare the type explicitly, so
that there is less chance of inadvertantly ending up with the wrong one.