Warn on service failure, debug stack trace
When a service fails to load in the ServiceHelper, it now issues a one
line warning message instead of an "error" with a full stack trace. The
stack trace is instead logged to debug.

The goal was to reduce the standard output, and avoid confusion over the
error message as program execution isn't actually halted by failing to
load a service.