ServiceHelper: fall back to stderr if there is no LogService yet
At the stage of the ServiceHelper, we cannot realistically assume
that the LogService has been created yet: there might have been an
error initializing it, it might not have been initialized yet, or
there might not be any LogService available to begin with.

We should fail gracefully in all those cases and simply fall back to
printing to stderr instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>