Make it possible to use a different PluginFinder
Previously, the Context always instantiated a PluginIndex which always
instantiated a DefaultPluginFinder and invoked it. This effectively
forced the usage of SezPoz with the system class loader.

When constructing a Context, it is now possible to explicitly give a
PluginIndex with its own explicitly given PluginFinder instance. In this
way, SezPoz discovery can be completely avoided if desired, or a
non-default ClassLoader can be given to the DefaultPluginFinder.