Give ReadInto a chance to return gently upon EOF
Earlier, we tried to make ReadInto interruptible by avoiding the
blocking readLine() call (instead calling ready() in a while loop that
slept for some time to give the Thread a chance to be interrupted).

This is a problem, though, because ready() will not return true if
the stream was closed, therefore readLine() cannot return null to
indicate EOF, and the Thread will be stuck in an infinite loop.

Instead, call readLine() directly and instead of trying to interrupt
the thread, close the input reader. That way, the loop will be
naturally interrupted upon EOF.

To keep the Thread interruptible even if readLine() blocks on a stream
that is not closed yet, close the stream in the interrupt() method.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>