ServiceHelper: do not report problem twice
Originally, we had a warn() and a debug() (the latter including the
stacktrace) when a service could not be instantiated. But the warning
was not helpful enough, so we added the stacktrace to the warning,
too. But then the exact same things were reported by both calls,
so let's remove the debug() call.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>