ServiceHelper: warn loudly when *no* classes were found
Most likely it is a serious error when no classes were found, and even
more likely it is the context class loader which is not set correctly.

Let's issue a helpful warning in that case (because this developer
spent way too much time debugging this very same issue not only one,
not only two, but three times in total! Well, no more!).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>