PluginInfo: allow specifying a custom class loader
It is now possible to pass a class loader explicitly to the PluginInfo
constructor, which will later be used when lazily loading the class.

This is important because it is not enough to use the default class
loader, or even the current thread's context class loader at the time of
lazy instantiation. Instead, we need the class loader that *was* the
relevant one at the time the class name was discovered (i.e., at the
time the PluginInfo was created).