DefaultPluginFinder: always specify the class loader
Previously, in the default case we relied on SezPoz to choose an
appropriate class loader. While it did make an appropriate choice (the
current thread's context class loader), we did not explicitly have a
guarantee that that's what it did. So instead, we explicitly manage our
own class loader instance so that we can always hand it off to the new
PluginInfo instances for later use. In this way, if the context class
loader later changes, the PluginInfo won't care, and will still do the
right thing at the time PluginInfo#loadClass() is called.