MirrorWebsite: prepare for file:// URLs
With http:// URLs, we relied on the web server to generate directory
listings if index.html is missing. With file:// URLs, nobody helps us,
we have to list the directory items ourselves.

This is imposed on us by inefficient bureaucracy requiring us to
download a 1.7G file every single night. Although: we would need this
code also if they were offering us an rsync facility.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>