MirrorWebsite: make job scheduling less fragile
Up to now, we had an elaborate, clever, and hence pretty fragile way
to schedule mirror jobs. As a consequence, things broke down when we
tried it on file:// URLs.

Instead of pretending that we're clever, let's use a simple (and
therefore robust) way to schedule jobs.

And sure enough, the executor service gets shut down properly now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>