Help SezPoz a little
SezPoz uses the current Thread's context class loader to discover our
plugins. But if the DefaultPluginFinder cannot be loaded by said class
loader, it is safe to assume that SezPoz cannot find the plugins wanted
by the caller, either.

So let's not use the current Thread's context class loader in that case,
but the best we can do in that case: the class loader that found the
DefaultPluginFinder.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>