Add HasPluginInfo "duck-type" interface
This interface is very similar to Contextual and Prioritized, in that it
provides optional functionality which will be automatically functional
if a PluginService#createInstance(s) method is used to instantiate the
plugin. The interface provides an explicit linkage between the plugin
instance and its metadata, which would otherwise be unavailable without
an O(n) search (where n is the number of plugins of that type).