SortablePlugin: implement HasPluginInfo as well
Since SortablePlugin is the primary convenient extension point for
plugin type implementations, it makes sense for it to be as rich as
possible. There is little harm in retaining access to a plugin's
associated PluginInfo metadata.

As always when adding new API to a type hierarchy, while consumers of
that API are generally OK, implementors of said API may run into
trouble. In this case, SortablePlugin implements both methods of
HasPluginInfo, but there can still be method clashes. Specifically,
ImageJ2's Tool interface already had a getInfo method with incompatible
method argument, so updating to the next minor version release of
scijava-common will require corresponding changes to that interface.