Update version properties to never use SNAPSHOTs
For build reproducibility, it is best for projects not to use SNAPSHOT
linkages. Of course, you can always use a SNAPSHOT if you need it by
overriding the relevant property in your own child POM(s). But let's
not facilitate usage of SNAPSHOT dependencies by default.

Note that some of the release versions in this commit do not yet exist
as of this writing, but these releases are all forthcoming.