Manage event handler registration explicitly
There is now a dedicated Service method, registerEventHandlers, that
subscribes the event handlers (by delegating upstream to
AbstractContextual). The reason is: as things stood, it was required to
always call super.initialize() in Service implementations to ensure that
event handlers were registered. This is too much of a burden on third
party programmers, resulting in boilerplate code that is easily
forgotten, and hence mysterious service failures.

Since the ServiceHelper takes care of calling Service#initialize(), it
may as well also call Service#registerEventHandlers() explicitly
afterwards, to ensure that the event handlers are *always* registered.

The end result is that downstream code generally no longer needs to call
super.initialize(), which is much simpler and less error prone.