Skip CheckSezpozTest when running in a JRE
It is really a shame, but it seems that all of the following strategies
fail in Eclipse:

- try to figure out where a JDK might be (there might be none)

- try to figure out where Eclipse is and use its APT processing (first,
  we cannot really figure out where Eclipse is when running or debugging
  regular Java classes with a main() method -- with JUnit tests, we could
  at least figure out where Eclipse's JUnit support classes are and hence
  where the current Eclipse is -- and there is no platform-independent
  Java way to enquire the executable path of the parent process of the
  current JVM; second, even then we would have to deal with tons of
  Eclipse-specific things we would have to provide to call Eclipse's APT)

- run Sezpoz' Indexer6 directly (it cannot be run directly, you have to have
  a RoundEnvironment and provide scanners for annotations and classes and
  mirror classes)

The most promising solution to the problem that Eclipse is still unable to
conform to the Java Compiler specification, at this time and age, would have
been to call the Indexer6 directly.

But this developer thinks that the effort required to basically replicate
the complete APT functionality just to run Sezpoz is not worth it: we could
instead tell the developer to get their act together and install a JDK
already (which they need anyway because Eclipse is *not* the gold standard,
the JDK is, so if things compile/run correctly in Eclipse, you *still* have
to test with plain JDK).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>