SortablePlugin: use title as string representation
The title is the most human-friendly string available from the
PluginInfo, so it makes sense to provide a default toString
implementation that uses it.

This change is especially useful now that singleton plugins are
automatically added to the ObjectService: you can get the list of
singletons back for a given plugin interface, and then display them as
human-friendly strings thanks to this new toString() implementation.