AbstractContextual does not need to hold references to event subscribers
We used to need that because the EventService would not hold references
itself. But we can easily hold a weak mapping between the event
handler objects and their proxy subscriber objects, which we do now.

This also gets rid of the finalize() method in all classes based on
AbstractContextual, which is a good thing!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>