Add tests for @Parameter injection mechanism
We want to make sure that *any* class which has @Parameters of type
org.scijava.service.Service or org.scijava.Context has those parameters
populated with the context's associated values, when
Context#inject(Object) is called on it. We also want to make sure that
*if* the object implements the Contextual interface, the result is
exactly the same when calling Contextual#setContext(Context).