ServiceHelper: change how service deps are filled
The context construction is now more aggressive about failing when
required services are not able to be filled.

A service can be marked optional by implementing the Optional interface,
in which case the context's service population will continue even if
that service cannot be filled.

For transitive service dependencies, if the dependency is marked with
"@Parameter(required = false)" then it is considered an optional
dependency which also will not halt service population with an
exception, but rather merely be set to null.

Otherwise, if a required service cannot be filled,
IllegalArgumentException is thrown. In the case of a transitive
dependency failure, the IllegalArgumentException will have a chain of
causes that clearly identify the ultimate culprit, as well as the
dependency chain which triggered the failure.

In all cases, invalid services are still logged as warnings, with a
stack trace when in debug mode.