AbstractGateway: make noargs ctor work again
When the default constructor is used, it passes a null context to the
other constructor, which then tries to call context.inject(this),
resulting in a NullPointerException. With this fix, you can call the
default constructor and end up with a null context. You'll still get an
NPE if you subsequently call any of the Gateway API methods before
calling setContext manually, though.