AbstractContextual: do not allow setContext(null)
This reverts commit 59a1b5be1180be4fcae21f8c4a48aefc98c549bb.

First of all, scifio 0.4.1 was pinned to scijava-common 1.6.1, not
2.0.0. So the information in the reverted commit above is wrong: SCIFIO
does use the trick of calling setContext(null) in a couple of places,
but this only worked in scijava-common 1.x, not scijava-common 2.0.0.

After consultation with Johannes, we decided that we would like
scijava-common 2.x to continue throwing an exception when
setContext(null) is called, rather than silently allowing it like
scijava-common 1.x did. This intentionally breaks backward compatibility
(e.g., scifio 0.4.1 cannot work with scijava-common 2.x). The rationale
is that we want to avoid the techical debt (i.e., future problems and
confusion) of setContext(null) being anything besides a logic error.

In short: setContext(null) is a bad idea, so let's not allow it.

The SCIFIO 0.5.0 release will no longer call setContext(null):
https://github.com/scifio/scifio/commit/d408857862d1a1b48d67cf1dd98ad87b3f182f47