Deprecate FileUtils#matchVersionedFilename
The only valid use we have for it is really to strip out versions
from versioned file names, but we have a better method now:
FileUtils#stripFilenameVersion.

That version is also forward-compatible because the regular expression
behind matchVersionedFilename is about to introduce new groups (and
therefore group(5) will not reliably match the extension).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>