CheckSezpoz: write APT properties more strictly
This is done because Eclipse 4.3+ removes comments automatically
(particularly the date comment that java.util.Properties#store always
prepends when writing its output). We also want to write the properties
in a very particular order, to minimize the chance of modifications.

See the big comment in this commit for further details.