CheckSezpoz: use better var for APT settings file
There is actually an "org.eclipse.jdt.core.prefs" file in addition to
"org.eclipse.jdt.apt.core.prefs" so it is ambiguous which one a variable
called "jdtSettings" would refer to. The name "aptSettings" is more
clear because AFAIK there is only one configuration file with the text
"apt" in the name, and indeed this configuration relates to the
annotation processing specifically.