CheckSezpoz: restrict visibility of methods
Methods marked "protected static" are accessible only A) from
subclasses; and B) from other classes in the same package.

Since CheckSezpoz is a static utility class, there will be no
subclasses, so "protected" is actually confusing. We instead use
private visibility for methods not used anywhere else, and default
(i.e., package) visibility for methods which need to be accessible
from CheckSezpozTest (which is in the same package).