Created ObjectArray
Added a new PrimitiveArray implementation for Objects. This is, of
course, a misnomer as Objects are not primitives.

However, it was necessary to create an array type ([]) wrapper, and it
makes sense to reuse the existing PrimitiveArray pattern.

It may make sense to rename PrimitiveArray to something like
"ArrayTypeWrapper", etc... but for now we focus on providing the new
Object wrapping functionality.