ConversionUtilsTest: handle private fields too
The ClassUtils#getField(Class, String) method uses getDeclaredField
rather than getField to obtain the name of the field, so that private
fields can also be accessed. Of course, then no subclasses are handled,
but for the purposes of this test class, it doesn't matter.

This change also lets us stop propagating NoSuchFieldException
throughout the tests, too.