ConversionUtils: fix bug in getComponentClass
It is not safe to cast directly to Class<?> because
getGenericComponentType() may return a non-Class Type. So we filter
through the getClass(Type) method to ensure the result is a Class.

We probably want to add a getComponentType(Type) method that returns a
Type, and deprecate getComponentClass(Type) in favor of downstream
callers chaining getClass(getComponentType(type)) explicitly.