WIP: Ideas for expanding the EventHistory API
Probably not necessary; the EventHistory is just an example of how to
subscribe to SciJavaEvents and do something with them. Typically you
will need to listen to a specific type of event and process it more
precisely than simply caching it to a string.

The current patch here is inconsistent about whether filtering should
include sub-event-types or not...