Accelerator: refine meaning of the caret (^)
On OS X, it was previously impossible to define a plugin's shortcut as
"control KEY" for any key. The control was always replaced with meta.
Ultimately, I consider this limitation to be a bug.

To fix it, the caret character now expands to control on non-Mac
platforms, and expands to meta (i.e., the command key) on Mac platforms.
99% of the time, this is the desired behavior for keyboard shortcuts.

However, if you actually want a plugin to always use the control key
across all platforms, you can now spell it out by writing "control" or
"ctrl" and it will no longer be replaced by meta on OS X.

Technically, this is a backwards incompatible change in behavior which
one could argue warrants a bump to 3.0.0 according to SemVer. However:

* It is not a backwards-incompatible change to the API proper.
* It is unlikely that any downstream code other than ImageJ2 will be
  affected by the change.
* The fix is trivial (replace spelled out "control " and "ctrl " strings
  with "^" in affected plugins).
* The previous behavior was inflexible and undesirable.

So we'll leave this as a "bug-fix" instead.