Add a helper for Eclipse
Eclipse is a bit challenged and challenging when it comes to following
the Java standard. For example, in the default mode Eclipse does not run
any annotation processors (as it would be mandated by the Java
specification).

To compensate for this, we simply run a custom annotation processor
whenever we encounter at Index.load() time to handle the case that there
might be some annotation index that is out-of-date.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>