Index annotations even when running tests in Maven
There is an unfortunate interaction between Eclipse and Maven when the
developer runs the tests using Maven after Eclipse compiled code, failing
to run the annotation processors. The annotation index breaks down in that
case because no annotations got indexed.

Fix this by running the index annotation even when the tests are run from
Maven's surefire plugin (as opposed to from within Eclipse).

This will help us also when merging the annotation indexer into
scijava-common itself, because scijava-common has annotations itself that
need to be indexed, but after the merge, scijava-common itself is the
annotation processor (which needs to be disabled in order to compile
itself because of a chicken-and-egg problem).

While at it, remember which directories we already looked at (and clear
that cache when running the EclipseHelper unit test).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>