EclipseHelper: use a different tell-tale than ':' for jar: URLs
When running the EclipseHelper to index annotations on directories, we
used to look for colons in the path to determine whether it might not
be a plain file. This interferes with those pesky DOS-style paths that
Windows still retains, though.

So just switch to look for a leading slash (which is used in URLs even on
Windows).

Pointed out by Jamion Jenkins.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>