Context: print more information on failed inject
If a required service is not found, we now print more information about
the context and classpath to assist the user in identifying what went
wrong.

Specifically, the list of services known to the context are printed out,
as well as the current thread's classpath's URLs.

Hopefully this will make failed service issues a little easier to
understand.