Split out pom-scijava into its own repository
So far, pom-scijava has lived in the same repository as scijava-common
just out of convenience: all the scripts to work with it lived in the
same repository, too.

Now that the scripts are gone into their own repository, it is time to
let pom-scijava live in its own repository, too:
https://github.com/scijava/pom-scijava

The commit history was rewritten in that repository so that this
commit's parent commit corresponds to pom-scijava's 8823626f(Increase
pom-scijava version to 1.119) commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>