FileUtils#appendContents: exclude directory itself
When listing the contents of a directory, we do not want the directory
itself to be in the returned collection. First of all, it is
inconsistent with the idea of listing the directory's "contents" (i.e.,
its children). And secondly, it makes it more involved to invoke the
method recursively.