Contextual: add non-null context accessor
The new context() method behaves like getContext(), except that if the
context is null, it throws a NullContextException (unchecked type).

The rationale is that it is extremely common to chain method calls like
"myObject.getContext().getService..." or similar. But if getContext()
returns null, that invocation throws NullPointerException. We would like
a slightly more informative exception to be thrown in that case.