Revert "Prettify recent change"
This reverts commit 98c2c13abc6c8e8be87c00449bf47092948662af.

Rationale: We decided to pass null down the line to the lower-level
getBaseDirectory(File, String) method in order to make the code more
future-proof. Should the getBaseDirectory(File, String) call be able to
do something sensible with a null File reference in the future (rather
than only returning null as it does now), then no further change to
getBaseDirectory(Class, String) will be necessary to accommodate that.
Conversely, with the short-circuit immediate return of null, this method
would need to be changed again to take advantage of any later
improvements to getBaseDirectory(File, String).