Work around problem with Batik on the class path
Batik unfortunately bundles xalan inside the .jar. As if this was not bad
enough, it does so with a real old version, one that is incompatible with
xalan we use. The symptom is a NoSuchMethodError stating that
org.apache.xpath.XPathContext.<init>(Z)V could not be found.

Let's work around that issue by simply ignoring it and continue on the
assumption that we cannot extract anything from the XML.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>