Fix EclipseHelper regression
EclipseHelper needs to fix the problem that Eclipse fails to run
annotation processors when compiling its classes. It not only needs to
do that when running inside Eclipse, unfortunately: When Eclipse
compiled the classes and Maven detects that it does not need to compile
any .java files, no annotation processors are run, either. Therefore,
the EclipseHelper needs to try harder even when we detect that it is
running from inside the maven-surefire-plugin instead of Eclipse.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>