Only write META-INF/json/ if annotations were found
Originally, we wanted to run the EclipseHelper only for a limited set of
situations, using the test whether the first class path element of the
current class loader is a directory as an indicator that we're running
in Eclipse. We also wanted to use the time stamp of META-INF/json/ as an
indicator whether we need to index the annotations or not.

However, it turns out that we probably need to run in more situations,
e.g. because the directories in question might be first indexed during a
run inside a LegacyEnvironment.

So let's just punt and run the EclipseHelper in more situations, but
avoid to write META-INF/json/ unless we really have anything to write.

This addresses https://github.com/scijava/scijava-common/issues/33.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>