ClassUtils: add methods for precise field typing
These methods exploit generics to deduce more details of the type of the
field in question, with respect to a particular (usually sub-)class.
See the method javadoc for further details.

This introduces an (encapsulated) dependency on the excellent gentyref
library, which does the bulk of the work.

Many thanks to Wouter Coekaerts for creating gentyref, and for his
patience in explaining some of the finer points of Java generics:
    https://github.com/coekie/gentyref/pull/2