Clarify docs for canConvert(Object, Class)
It was misleading the way it was because it would return true from
canConvert("5.1", int.class) but would get an exception when actually
trying to convert("5.1", int.class). The documentation led me to believe
the check would catch this scenario when it is actually a convenience
method that simply delegates to canConvert(Class, Class).