AbstractInputHarvester: fix unsafe widget cast
We now explicitly check if the widget is the right type before using it.

In practice, the AbstractInputPanel#supports(InputWidget) method already
checked this, so it should never be the case that an incompatible
widget is returned by WidgetService#create(WidgetModel). But there is no
guarantee that all downstream code is using the abstract superclasses,
so let's check again -- better safe than sorry here.