EclipseHelper: skip unpacked ImageJ1 classes
This is a hacky workaround for the situation where ImageJ1 classes got
loaded prematurely by the Eclipse helper when running from within
Eclipse with a project coupling to ImageJA. Since we know that ImageJ1
will never use SciJava plugin annotations, we can safely skip loading
those classes, thus preventing the premature loading in that scenario.

As usual, thanks to Johannes Schindelin for his help with this fix!