DirectoryIndexer: Handle Throwables
This skips classes which can't be loaded due to a SecurityException during
the annotation pre-processing using the Eclipse Helper. Reason: Not all
classes of a eclipse-plugin can be loaded outside the eclipse-enviroment,
i.e. from a "normal" java program. If you try to load them during
annotation-preprocessing using the EclipseHelper, the SecurityException
is thrown.